hot encoding changes

+++ framework-service/src/main/java/com/inferyx/framework/service/RunModelServiceImpl.java
@@ -57,6 +57,7 @@ import com.inferyx.framework.domain.Status;
 import com.inferyx.framework.domain.Train;
 import com.inferyx.framework.domain.TrainExec;
 import com.inferyx.framework.domain.TrainResult;
+import com.inferyx.framework.enums.EncodingType;
 import com.inferyx.framework.enums.RunMode;
 import com.inferyx.framework.executor.DL4JExecutor;
 import com.inferyx.framework.executor.ExecContext;
@@ -736,7 +737,7 @@ public class RunModelServiceImpl implements Callable<TaskHolder> {
 			String filePath = String.format("/%s/%s/%s", train.getUuid(), train.getVersion(), trainExec.getVersion());
 			String tableName = String.format("%s_%s_%s", train.getUuid().replace("-", "_"), train.getVersion(), trainExec.getVersion());
 			String appUuid = commonServiceImpl.getApp().getUuid();
-			
+
 			if (!modelType.equalsIgnoreCase(ExecContext.R.toString())
 					&& !modelType.equalsIgnoreCase(ExecContext.PYTHON.toString())) {
 				Algorithm algorithm = (Algorithm) commonServiceImpl.getOneByUuidAndVersion(algorithmUUID, algorithmVersion, MetaType.algorithm.toString(), "N");
@@ -746,6 +747,8 @@ public class RunModelServiceImpl implements Callable<TaskHolder> {
 
 				String filePathUrl = String.format("%s%s%s", Helper.getPropertyValue("framework.hdfs.URI"), Helper.getPropertyValue("framework.model.train.path"), filePath);
 				String testSetPath = filePathUrl.endsWith("/") ? filePathUrl.concat("test_set") : filePathUrl.concat("/").concat("test_set");
+				
+				Map<String, EncodingType> encodingDetails = getEncodingDetailsByFeatureAttrMap(train.getFeatureAttrMap());
 				
 				trainOtherParam.put("confusionMatrixTableName", trainName+"confusionMatrix");
 				exec = execFactory.getExecutor(datasource.getType());
@@ -758,8 +761,6 @@ public class RunModelServiceImpl implements Callable<TaskHolder> {
 //				exec.executeAndRegister(sql, (tableName+"_train_data"), appUuid);
 				
 				String sourceSql = modelServiceImpl.generateSQLBySource(source, execParams);
-				
-				
 				
 				ResultSetHolder sourceRsHolder = exec.executeAndRegisterByDatasource(sourceSql, (tableName+"_train_source_data"), trainSrcDatasource, appUuid);
 				sourceRsHolder.setTableName((tableName+"_train_source_data"));
@@ -793,7 +794,7 @@ public class RunModelServiceImpl implements Callable<TaskHolder> {
 					trndModel = exec.train(paramMap, fieldArray, label, algorithm.getTrainClass(), train.getTrainPercent()
 							, train.getValPercent(), (tableName+"_train_source_data"), tempTrngDfSql
 							, algoclass, trainOtherParam, trainResult, testSetPath
-							, rowIdentifierCols, train.getIncludeFeatures(), tempTrngDfSql, tempValDfSql);
+							, rowIdentifierCols, train.getIncludeFeatures(), tempTrngDfSql, tempValDfSql, encodingDetails);
 				} else if (!model.getType().equalsIgnoreCase(ExecContext.DL4J.toString())) {		
 					//With hypertuning
 					List<ParamListHolder> paramListHolderList = null;
@@ -807,7 +808,7 @@ public class RunModelServiceImpl implements Callable<TaskHolder> {
 									trndModel = exec.trainCrossValidation(paramMap, fieldArray, label, algorithm.getTrainClass()
 											, train.getTrainPercent(), train.getValPercent(), (tableName+"_train_source_data")
 											, hyperParamList.getParams(), appUuid, trainOtherParam, trainResult, testSetPath
-											, rowIdentifierCols, train.getIncludeFeatures(), tempTrngDfSql, tempValDfSql);
+											, rowIdentifierCols, train.getIncludeFeatures(), tempTrngDfSql, tempValDfSql, encodingDetails);
 								}
 							}
 						} else if(execParams.getParamListInfo() != null) {
@@ -819,7 +820,7 @@ public class RunModelServiceImpl implements Callable<TaskHolder> {
 								trndModel = exec.trainCrossValidation(paramMap, fieldArray, label, algorithm.getTrainClass()
 										, train.getTrainPercent(), train.getValPercent(), (tableName+"_train_source_data")
 										, hyperParamList.getParams(), appUuid, trainOtherParam, trainResult, testSetPath
-										, rowIdentifierCols, train.getIncludeFeatures(), tempTrngDfSql, tempValDfSql);
+										, rowIdentifierCols, train.getIncludeFeatures(), tempTrngDfSql, tempValDfSql, encodingDetails);
 							}
 						}
 					} else {
@@ -835,7 +836,7 @@ public class RunModelServiceImpl implements Callable<TaskHolder> {
 							trndModel = exec.trainCrossValidation(paramMap, fieldArray, label, algorithm.getTrainClass()
 									, train.getTrainPercent(), train.getValPercent(), (tableName+"_train_source_data")
 									, hyperParamList.getParams(), appUuid, trainOtherParam, trainResult, testSetPath
-									, rowIdentifierCols, train.getIncludeFeatures(), tempTrngDfSql, tempValDfSql);
+									, rowIdentifierCols, train.getIncludeFeatures(), tempTrngDfSql, tempValDfSql, encodingDetails);
 						}
 					}
 				} else {
@@ -853,29 +854,31 @@ public class RunModelServiceImpl implements Callable<TaskHolder> {
 				boolean isModelSved = modelServiceImpl.save(algorithm.getModelClass(), trndModel, filePathUrl+"/model");
 				String defaultDir = null;
 				if (algorithm.getSavePmml().equalsIgnoreCase("Y")) {
-					try {
-						String filePathUrl_2 = null;
-						if(filePathUrl.contains(hdfsInfo.getHdfsURL()))
-							filePathUrl_2 = filePathUrl.replaceAll(hdfsInfo.getHdfsURL(), "");
-						defaultDir = filePathUrl_2;
-						String pmmlLocation = filePathUrl_2 + "/" + train.getUuid() + "_" + train.getVersion() + "_"
-								+ (filePathUrl_2.substring(filePathUrl_2.lastIndexOf("/") + 1)) + ".pmml";
-						
-						File pmmlFile = new File(pmmlLocation);
-						if(pmmlFile.exists()) {
-						FileUtils.forceDelete(pmmlFile);
-						} 
-						boolean isSaved = exec.savePMML(trndModel, "trainedDataSet", pmmlLocation, appUuid);
-						if(isSaved) {
-							logger.info("PMML saved at location: "+pmmlLocation);
-						} else {
-							logger.info("PMML not saved.");
+					if(encodingDetails == null || (encodingDetails != null && encodingDetails.isEmpty())) {
+						try {
+							String filePathUrl_2 = null;
+							if(filePathUrl.contains(hdfsInfo.getHdfsURL()))
+								filePathUrl_2 = filePathUrl.replaceAll(hdfsInfo.getHdfsURL(), "");
+							defaultDir = filePathUrl_2;
+							String pmmlLocation = filePathUrl_2 + "/" + train.getUuid() + "_" + train.getVersion() + "_"
+									+ (filePathUrl_2.substring(filePathUrl_2.lastIndexOf("/") + 1)) + ".pmml";
+							
+							File pmmlFile = new File(pmmlLocation);
+							if(pmmlFile.exists()) {
+							FileUtils.forceDelete(pmmlFile);
+							} 
+							boolean isSaved = exec.savePMML(trndModel, "trainedDataSet", pmmlLocation, appUuid);
+							if(isSaved) {
+								logger.info("PMML saved at location: "+pmmlLocation);
+							} else {
+								logger.info("PMML not saved.");
+							}
+						 }catch (JAXBException e) {
+							e.printStackTrace();
+						} catch (Exception e) {
+							e.printStackTrace();
 						}
-					 }catch (JAXBException e) {
-						e.printStackTrace();
-					} catch (Exception e) {
-						e.printStackTrace();
-					}
+					}					
 				}
 				if (isModelSved) {	
 					defaultDir = filePathUrl.replaceAll(hdfsInfo.getHdfsURL(), "");
@@ -884,26 +887,33 @@ public class RunModelServiceImpl implements Callable<TaskHolder> {
 						Map<String, Object> summary = exec.summary(trndModel, algorithm.getSummaryMethods(), appUuid);
 						
 						String fileName = tableName+".result";
-						summary = exec.calculateConfusionMatrixAndRoc(summary,trainOtherParam.get("confusionMatrixTableName"),appUuid);
-						
-						double[] featureimportancesArr = (double[])summary.get("featureimportances");						
-						if(featureimportancesArr != null) {
-							List<Double> featureImportance = Arrays.asList(ArrayUtils.toObject(featureimportancesArr));
-							trainResult.setFeatureImportance(featureImportance);
+						if(encodingDetails == null || (encodingDetails != null && encodingDetails.isEmpty())) {
+							summary = exec.calculateConfusionMatrixAndRoc(summary,trainOtherParam.get("confusionMatrixTableName"),appUuid);
 						}
-						trainResult.setAccuracy((double) summary.get("accuracy"));
-						trainResult.setRecall((double) summary.get("recall"));
-						trainResult.setF1Score((double) summary.get("f1Score"));
-						trainResult.setPrecision((double) summary.get("precision"));
-//						trainResult.setRocAUC((List<Double>) summary.get("roc"));
 						
-						writeSummaryToFile(summary, defaultDir, fileName);
+						if(summary != null && !summary.isEmpty()) {
+							double[] featureimportancesArr = (double[])summary.get("featureimportances");						
+							if(featureimportancesArr != null) {
+								List<Double> featureImportance = Arrays.asList(ArrayUtils.toObject(featureimportancesArr));
+								trainResult.setFeatureImportance(featureImportance);
+							}
+							trainResult.setAccuracy((double) summary.get("accuracy"));
+							trainResult.setRecall((double) summary.get("recall"));
+							trainResult.setF1Score((double) summary.get("f1Score"));
+							trainResult.setPrecision((double) summary.get("precision"));
+//							trainResult.setRocAUC((List<Double>) summary.get("roc"));
+							
+							writeSummaryToFile(summary, defaultDir, fileName);
+						}
 					} else if(trndModel instanceof PipelineModel) {
 						filePathUrl = filePathUrl+"/model" + "/stages/" + customDirectories.get(1) + "/data/";
 						Map<String, Object> summary = exec.summary(trndModel, algorithm.getSummaryMethods(), appUuid);
 												
 						String fileName = tableName+".result";
-						summary = exec.calculateConfusionMatrixAndRoc(summary,trainOtherParam.get("confusionMatrixTableName"),appUuid);
+
+						if(encodingDetails == null || (encodingDetails != null && encodingDetails.isEmpty())) {
+							summary = exec.calculateConfusionMatrixAndRoc(summary,trainOtherParam.get("confusionMatrixTableName"),appUuid);
+						}
 						double[] featureimportancesArr = (double[])summary.get("featureimportances");	
 						if(featureimportancesArr != null) {
 							List<Double> featureImportance = Arrays.asList(ArrayUtils.toObject(featureimportancesArr));
@@ -1000,6 +1010,20 @@ public class RunModelServiceImpl implements Callable<TaskHolder> {
 		}		
 	}
 	
+	private Map<String, EncodingType> getEncodingDetailsByFeatureAttrMap(List<FeatureAttrMap> featureAttrMap) {
+		Map<String, EncodingType> encodingDetails = new LinkedHashMap<>();
+		for(FeatureAttrMap attrMap : featureAttrMap) {
+			if(attrMap.getEncodingType() != null) {
+				encodingDetails.put(attrMap.getFeature().getFeatureName(), attrMap.getEncodingType());
+			}
+		}
+		if(!encodingDetails.isEmpty()) {
+			return encodingDetails;
+		}else {
+			return null;
+		}
+	}
+
 	public String writeSummaryToFile(Map<String, Object> summary, String directory, String fileName) throws IOException {		
 		String filePath = directory + "/" + fileName;		
 		String resultJson = new ObjectMapper().writeValueAsString(summary);
